pyproject.toml
[project]
name = "datadup"
version = "0.1.0"
description = "Check if your test set is contained in train"
authors = [
    {name = "Justin T Chiu", email = "chiu.justin.t@gmail.com"},
]
dependencies = [
    "datasets",
    "transformers",
    "numpy",
    "scipy",
    "datasketch",
    "dpu-utils",
    "tqdm",
    "mosaicml-streaming",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = [
    "pdm-backend",
    "numpy",
    "setuptools",
]
build-backend = "pdm.backend"
